# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FpolyZone",
        "version": "0.1"
    },
    "paths": {
...
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Đăng ký tài khoản",
                "description": "Tạo một tài khoản mới.",
                "operationId": "d764dd091cc4494ae0baf360b03319f3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "username",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "username": {
                                        "description": "Tên đăng nhập của người dùng",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Địa chỉ email của người dùng",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "description": "Mật khẩu của người dùng",
                                        "type": "string",
                                        "minLength": 8
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tạo tài khoản thành công"
                    },
                    "400": {
                        "description": "Lỗi trong quá trình xử lý"
                    },
                    "422": {
                        "description": "Dữ liệu không hợp lệ"
                    }
                }
            }
        },
        "/api/auth/verify": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Xác minh tài khoản",
                "description": "Xác minh tài khoản người dùng bằng mã xác nhận.",
                "operationId": "86a3349a0306cb26904691a7211faf38",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "verification_code"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Địa chỉ email của người dùng",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "verification_code": {
                                        "description": "Mã xác nhận",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Xác minh tài khoản thành công"
                    },
                    "403": {
                        "description": "Mã xác nhận không chính xác"
                    },
                    "404": {
                        "description": "Không tìm thấy tài khoản"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Đăng nhập",
                "description": "Đăng nhập vào hệ thống bằng địa chỉ email và mật khẩu.",
                "operationId": "8dcb70df1020986038d098cc08d05dae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Địa chỉ email của người dùng",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "description": "Mật khẩu của người dùng",
                                        "type": "string",
                                        "minLength": 8
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Đăng nhập thành công"
                    },
                    "400": {
                        "description": "Đăng nhập thất bại"
                    },
                    "403": {
                        "description": "Tài khoản chưa được kích hoạt"
                    },
                    "422": {
                        "description": "Dữ liệu không hợp lệ"
                    }
                }
            }
        },
...
        "/api/auth/google-callback": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Xử lý đăng nhập bằng Google",
                "description": "Xử lý việc đăng nhập bằng Google, lấy thông tin trả về từ Google thực hiện đăng ký hoặc đăng nhập.",
                "operationId": "efce55f6e0a2c532daa7c439214d443c",
                "responses": {
                    "200": {
                        "description": "Đăng nhập bằng Google thành công"
                    },
                    "403": {
                        "description": "Tài khoản bị khóa hoặc lỗi trong quá trình xử lý"
                    },
                    "400": {
                        "description": "Lỗi trong quá trình xử lý"
                    }
                }
            }
        }
    }
}